Split `board` out from `game`
