+++ src/board/StatusBar.js
