+++ src/board/index.ts
@@ -1,0 +1,7 @@
+export { default as MachikoroBoard } from './Board';
+export { default as Buttons } from './Buttons';
+export { default as Establishments } from './Establishments';
+export { default as Log } from './Log';
+export { default as PlayerInfo } from './PlayerInfo';
+export { default as StackTable } from './StackTable';
+export { default as StatusBar } from './StatusBar';
