+++ src/game/index.ts
@@ -1,0 +1,2 @@
+export * from './Game';
+export * from './meta';
